# api/openapi/fleet-v1.yaml
...
openapi: 3.1.0
info:
  title: agent-fleet control plane
  version: 0.1.0
  description: |
    面向单操作者的 Agent 工作站机群管理器控制面。
    资源统一 K8s 风格 `metadata/spec/status` JSON 形态；名称在类型内唯一。
    spec 由操作者读写；status 由控制器维护——创建时初始化为 `{}`，更新时不接受改写。
servers:
  - url: http://127.0.0.1:7788
# §29.14：admin token 经 `Authorization: Bearer` 传递；服务端未配置 token（回环默认）
# 时不校验。因此 security 列表含空项——"带 token"与"服务端未启用鉴权"都合法。
security:
  - adminToken: []
  - {}
...
paths:
...
  /api/v1/machines:
    get:
      tags: [machines]
      summary: 列出 Machine（支持 ?name= 过滤）
      parameters:
        - name: name
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Machine 列表
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MachineList"
    post:
      tags: [machines]
      summary: 创建 Machine
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Machine"
      responses:
        "201":
          description: 已创建（服务端补齐 metadata，status 置 {}）
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Machine"
        "400": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
  /api/v1/machines/{name}:
    parameters:
      - $ref: "#/components/parameters/ResourceName"
    get:
      tags: [machines]
      summary: 读取 Machine
      responses:
        "200":
          description: Machine
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Machine"
        "404": { $ref: "#/components/responses/Error" }
    put:
      tags: [machines]
      summary: 更新 Machine spec（metadata 与 status 以存储现值为准）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Machine"
      responses:
        "200":
          description: 更新后的 Machine
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Machine"
        "400": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
    delete:
      tags: [machines]
      summary: 删除 Machine
      responses:
        "204":
          description: 已删除
        "404": { $ref: "#/components/responses/Error" }
  /api/v1/profiles:
    get:
      tags: [profiles]
      summary: 列出 AgentProfile（支持 ?name= 过滤）
      parameters:
        - name: name
          in: query
          schema:
            type: string
      responses:
        "200":
          description: AgentProfile 列表
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentProfileList"
    post:
      tags: [profiles]
      summary: 创建 AgentProfile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentProfile"
      responses:
        "201":
          description: 已创建
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentProfile"
        "400": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
  /api/v1/profiles/{name}:
    parameters:
      - $ref: "#/components/parameters/ResourceName"
    get:
      tags: [profiles]
      summary: 读取 AgentProfile
      responses:
        "200":
          description: AgentProfile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentProfile"
        "404": { $ref: "#/components/responses/Error" }
    put:
      tags: [profiles]
      summary: 更新 AgentProfile spec
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentProfile"
      responses:
        "200":
          description: 更新后的 AgentProfile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentProfile"
        "400": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
    delete:
      tags: [profiles]
      summary: 删除 AgentProfile
      responses:
        "204":
          description: 已删除
        "404": { $ref: "#/components/responses/Error" }
  /api/v1/machines/{name}/enroll-token:
    post:
      tags: [machines]
      summary: 签发一次性 enrollment token（KM-22）
      description: |
        为既有 Machine 签发 enrollment token（短时效、一次性、熵 ≥128-bit），
        并绑定 machineId 与 CSR 公钥指纹（§4.6）。token 明文仅在响应中出现一次，
        服务端只存 SHA-256 哈希；并发/重放的 enroll 消费在数据库层原子化，
        至多一个请求成功（FR-11.5）。
      parameters:
        - $ref: "#/components/parameters/ResourceName"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnrollTokenRequest"
      responses:
        "201":
          description: 已签发（含一次性明文 token）
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnrollToken"
        "400": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
  # ===== KM-23 动作端点（§8.1）=====
  /api/v1/machines/{name}/reconcile:
    post:
      tags: [machines]
      summary: 触发收敛或确认计划（KM-23）
      description: |
        机器级互斥的派发点获取（§4.3）：以一次事务创建 Pending 操作行，
        该机已有未决操作（Pending/Running/AwaitingConfirmation/CancelRequested/
        Unknown）时返回 409 MachineBusy（diagnostics 携带未决操作引用），不排队。
        携带 confirmPlanDigest 时为互斥三例外之一（§4.3 第 6 条）：确认该操作
        自己的计划并推进执行；digest 与等待中的 planDigest 不一致返回
        409 ReplanRequired（FR-12.7）。
      parameters:
        - $ref: "#/components/parameters/ResourceName"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReconcileRequest"
      responses:
        "202":
          description: 已受理（返回 Operation，客户端凭 id 追踪进度）
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Operation"
        "409": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
  /api/v1/machines/{name}/rollback:
    post:
      tags: [machines]
      summary: 机器级回滚到指定 generation（KM-23，§7.2）
      description: |
        对目标 generation 的快照重放一次 reconcile：先把目标代内容物化为该机
        新期望代（回滚不把机器拉回旧代，FR-7.5/ADR-2），走适配器合并写
        （天然保留未托管字段）。目标代不可解析 → 409 RollbackUnsupported。
      parameters:
        - $ref: "#/components/parameters/ResourceName"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [targetGeneration]
              properties:
                targetGeneration:
                  type: integer
                  format: int64
      responses:
        "202":
          description: 已受理（Rollback 型 Operation）
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Operation"
        "409": { $ref: "#/components/responses/Error" }
  /api/v1/machines/{name}/operations:
    get:
      tags: [machines]
      summary: 列出机器的操作审计记录（KM-23）
      parameters:
        - $ref: "#/components/parameters/ResourceName"
      responses:
        "200":
          description: 操作列表（§6.4 不可变审计单元）
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperationList"
        "404": { $ref: "#/components/responses/Error" }
  /api/v1/machines/{name}/operations/{opId}/cancel:
    post:
      tags: [machines]
      summary: 取消未决操作（互斥例外端点之一，KM-23）
      description: |
        置 CancelRequested 并投递 CancelOperation；节点不可达时保持
        CancelRequested，服务端不单方面置 Cancelled（FR-13.9）。重复取消幂等。
        AwaitingConfirmation 的计划取消直接终态（节点无可中断流水线，§9.3）。
      parameters:
        - $ref: "#/components/parameters/ResourceName"
        - name: opId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: 取消状态（Operation）
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Operation"
        "404": { $ref: "#/components/responses/Error" }
  /api/v1/machines/{name}/operations/{opId}/skip:
    post:
      tags: [machines]
      summary: 显式跳过未决操作（互斥例外端点之一，KM-23）
      description: |
        立即释放控制面互斥，terminalModifier=Skipped：不计入成功、必须携带
        reason（审计），UI 显式标出，不允许静默跳过（§9.6 路径 C/FR-15.4）。
        跳过不改变节点实际状态——节点可能仍在写。
      parameters:
        - $ref: "#/components/parameters/ResourceName"
        - name: opId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason: { type: string }
                operator: { type: string }
      responses:
        "200":
          description: 跳过后的 Operation（Succeeded + Skipped 修饰）
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Operation"
        "400": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
  /api/v1/machines/{name}/drift:
    get:
      tags: [machines]
      summary: drift 三态视图（KM-23，FR-14.5）
      description: |
        一致(False) / 未知或过期(Unknown + reason: NeverInventoried |
        StaleObservation | ObservationPredatesDesired | ProjectionVersionMismatch) /
        漂移(True)。权威判据：期望受管投影摘要 == 观测受管投影摘要 且
        canonicalizationVersion 一致（FR-8.6）。
      parameters:
        - $ref: "#/components/parameters/ResourceName"
      responses:
        "200":
          description: drift 三态与证据
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MachineDrift"
        "404": { $ref: "#/components/responses/Error" }
  /api/v1/profiles/{name}/render:
    get:
      tags: [profiles]
      summary: 渲染期望状态预览（KM-23，FR-7.4）
      parameters:
        - $ref: "#/components/parameters/ResourceName"
        - name: machine
          in: query
          required: true
          schema: { type: string }
      responses:
        "200":
          description: 归一化期望状态 + 确定性摘要（不落库、不增代）
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RenderPreview"
        "400": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
  /api/v1/deployments/{name}/rollback:
    post:
      tags: [machines]
      summary: 创建回滚型 Deployment（KM-23，FR-10.4）
      description: |
        新 Deployment 携 spec.rollbackOf = 被回滚 generation，指向历史
        targetGeneration（内容代）；历史不可变。推进时逐机物化新代。
      parameters:
        - name: name
          in: path
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                targetGeneration:
                  type: integer
                  format: int64
                  description: 缺省为源 Deployment 的目标代
      responses:
        "202":
          description: 回滚型 Deployment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deployment"
        "404": { $ref: "#/components/responses/Error" }
  /api/v1/deployments/{name}/targets/{machine}/skip:
    post:
      tags: [machines]
      summary: 显式跳过被阻塞的发布目标（KM-23，FR-10.7）
      parameters:
        - name: name
          in: path
          required: true
          schema: { type: string }
        - name: machine
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason: { type: string }
      responses:
        "200":
          description: 已跳过（记 Skipped + 审计；不计入成功）
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "404": { $ref: "#/components/responses/Error" }

  # ===== KM-25：skills / providers / deployments 的完整 CRUD（KM-23 起已实现）=====
  /api/v1/skills:
    get:
      tags: [skills]
      summary: 列出 Skill（支持 ?name= 过滤）
      parameters:
        - name: name
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Skill 列表
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SkillList"
    post:
      tags: [skills]
      summary: 创建 Skill
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Skill"
      responses:
        "201":
          description: 已创建
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Skill"
        "400": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
  /api/v1/skills/{name}:
    parameters:
      - $ref: "#/components/parameters/ResourceName"
    get:
      tags: [skills]
      summary: 读取 Skill
      responses:
        "200":
          description: Skill
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Skill"
        "404": { $ref: "#/components/responses/Error" }
    put:
      tags: [skills]
      summary: 更新 Skill spec（metadata 与 status 以存储现值为准）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Skill"
      responses:
        "200":
          description: 更新后的 Skill
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Skill"
        "400": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
    delete:
      tags: [skills]
      summary: 删除 Skill
      responses:
        "204":
          description: 已删除
        "404": { $ref: "#/components/responses/Error" }
  /api/v1/providers:
    get:
      tags: [providers]
      summary: 列出 ModelProvider（支持 ?name= 过滤；只存 apiKeyEnv 名称，不含秘密值）
      parameters:
        - name: name
          in: query
          schema: { type: string }
      responses:
        "200":
          description: ModelProvider 列表
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelProviderList"
    post:
      tags: [providers]
      summary: 创建 ModelProvider
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ModelProvider"
      responses:
        "201":
          description: 已创建
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelProvider"
        "400": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
  /api/v1/providers/{name}:
    parameters:
      - $ref: "#/components/parameters/ResourceName"
    get:
      tags: [providers]
      summary: 读取 ModelProvider
      responses:
        "200":
          description: ModelProvider
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelProvider"
        "404": { $ref: "#/components/responses/Error" }
    put:
      tags: [providers]
      summary: 更新 ModelProvider spec
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ModelProvider"
      responses:
        "200":
          description: 更新后的 ModelProvider
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelProvider"
        "400": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
    delete:
      tags: [providers]
      summary: 删除 ModelProvider
      responses:
        "204":
          description: 已删除
        "404": { $ref: "#/components/responses/Error" }
  /api/v1/deployments:
    get:
      tags: [deployments]
      summary: 列出 Deployment（支持 ?name= 过滤；status.targets 为逐机推进状态）
      description: |
        `status.targets` 由服务端从 `deployment_targets` 表补齐（§6.1 的资源形态、
        §10.1 的权威存储），列表与单对象、以及创建/更新/回滚的响应都带该字段；
        逐机结果的 `phase`/`reason` 是 UI 区分 Superseded/Skipped 与 Failed 的依据
        （FR-14.5 第 3 组，KM-25）。
      parameters:
        - name: name
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Deployment 列表
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeploymentList"
    post:
      tags: [deployments]
      summary: 创建 Deployment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Deployment"
      responses:
        "201":
          description: 已创建
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deployment"
        "400": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
  /api/v1/deployments/{name}:
    parameters:
      - name: name
        in: path
        required: true
        schema: { type: string }
    get:
      tags: [deployments]
      summary: 读取 Deployment
      responses:
        "200":
          description: Deployment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deployment"
        "404": { $ref: "#/components/responses/Error" }
    put:
      tags: [deployments]
      summary: 更新 Deployment spec
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Deployment"
      responses:
        "200":
          description: 更新后的 Deployment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deployment"
        "400": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
    delete:
      tags: [deployments]
      summary: 删除 Deployment
      responses:
        "204":
          description: 已删除
        "404": { $ref: "#/components/responses/Error" }

  # ===== SSE（KM-25；FR-14.1/FR-14.2，事件格式见 §23.6）=====
  /api/v1/events:
    get:
      tags: [events]
      summary: 订阅资源变更事件流（SSE）
      description: |
        `Accept: text/event-stream` 的长连接。每条事件为
        `event: <resource-type>` + `data: {"id","revision"}`：
        资源类型取 machines/profiles/skills/providers/deployments/operations，
        `id` 为资源名（operations 为操作 id，等于 metadata.name），
        `revision` 为写入后的 metadata.resourceVersion。
        UI 据此按 (类型, id) 选择性重取，SSE 不承载资源正文（FR-14.2/NFR-9）。

        契约细节（§23.6 未规定、本实现取的最小自洽语义，KM-25 已登记）：
        - 连接建立时先发一条标准 SSE `retry:` 建议（3s）与周期保活注释行；
        - 无 Last-Event-ID 重放：客户端重连后自行做一次全量重取（事件只用于失效旧数据）；
        - 删除事件携带"删除前版本 +1"，该 id 随后经 REST 读取得到 404；
        - 订阅端缓冲溢出时服务端关闭该连接（不静默丢事件），由客户端重连收敛；
        - 鉴权同其余 /api/v1 端点（§29.14，Bearer adminToken）。浏览器 EventSource
          无法设置请求头，UI 侧用 fetch 流式读取该端点。
      responses:
        "200":
          description: 事件流（text/event-stream；连接保持直到客户端断开）
          content:
            text/event-stream:
              schema:
                type: string
                description: |
                  SSE 记录序列，形如：
                  `event: machines\ndata: {"id":"gpu-home","revision":7}\n\n`
        "401": { $ref: "#/components/responses/Error" }
        "501":
          description: 未装配事件枢纽（骨架语义）
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

# 【KM-25 修复】原文件缺少顶层 `components:` 键：parameters/schemas/pathItems/responses
# 实际落在文档根上，全部 `#/components/...` 引用都解析不到，且两个 `schemas:` 同级键
# 在解析时后者覆盖前者（EnrollTokenRequest/EnrollToken 丢失）。现补齐 components 容器。
components:
  securitySchemes:
    adminToken:
      type: http
      scheme: bearer
      description: |
        §29.14 管理令牌：非回环监听时必须携带（值来自 `--admin-token-env` 指定的
        环境变量）；回环监听且未配置时不校验。浏览器端 EventSource 不能带头，
        UI 对 `GET /api/v1/events` 使用 fetch 流式读取以携带该头。
  parameters:
    ResourceName:
      name: name
      in: path
      required: true
      schema:
        type: string
        pattern: "^[A-Za-z0-9]([A-Za-z0-9._-]{0,251}[A-Za-z0-9])?$"
  responses:
    Error:
      description: 统一错误体
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  # 【KM-25 修复】原先 components 下出现两个 `schemas:` 键：YAML 解析时后者覆盖前者，
  # 使 EnrollTokenRequest / EnrollToken 在解析结果中消失（严格校验器也会拒绝重复键）。
  # 现合并为单一 schemas 块；原先未被任何路径引用的 pathItems.Skeleton 已删除
  # （骨架端点属第 6 片，未注册即无契约可写，见文件头说明）。
  schemas:
...
    ReconcileRequest:
      type: object
      description: |
        POST /machines/{name}/reconcile 的请求体（可空）。携带 confirmPlanDigest 时
        表示"确认该操作自己的计划"——机器级互斥三例外之一（§4.3 第 6 条），
        不创建第二条未决操作；digest 与等待中的 planDigest 不一致返回 409 ReplanRequired。
      properties:
        confirmPlanDigest:
          type: string
          description: 被确认计划（spec.planDigest）的摘要
    EnrollTokenRequest:
      type: object
      required: [csrPubKeySha256]
      properties:
        csrPubKeySha256:
          type: string
          description: agentd 生成的 CSR 公钥指纹（enroll 输出）；token 与其绑定
          pattern: "^sha256:[0-9a-f]{64}$"
          examples: ["sha256:1f38..."]
        ttlSeconds:
          type: integer
          description: 可选有效期（秒）；默认 600（§4.6 默认 10 分钟）
    EnrollToken:
      type: object
      required: [machineId, token, expiresAt]
      properties:
        machineId:
          type: string
        token:
          type: string
          description: 一次性明文 token；仅出现一次
        expiresAt:
          type: string
          format: date-time
        note:
          type: string
...
    Machine:
      type: object
      required: [metadata]
      properties:
        metadata:
          $ref: "#/components/schemas/ObjectMeta"
        spec:
          type: object
          description: |
            Machine 规格（JSON 对象透传存储）。已知字段：
            managementMode（agentd|ssh，缺省 agentd）、profileRef、ssh。
          properties:
            managementMode:
              type: string
              enum: [agentd, ssh]
            profileRef:
              type: string
          additionalProperties: true
        status:
          description: 控制器维护（§6.2 conditions 与观测字段）；API 创建/更新时不接受改写
          $ref: "#/components/schemas/MachineStatus"
    MachineList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Machine"
...
    Operation:
      type: object
      description: 系统创建的不可变审计单元（§6.4）；动作端点以 202 返回
      required: [metadata, spec, status]
      properties:
        metadata: { $ref: "#/components/schemas/ObjectMeta" }
        spec:
          type: object
          properties:
            machine: { type: string }
            type:
              type: string
              enum: [Reconcile, Repair, Bootstrap, Rollback, AutoPlan]
            transport: { type: string, enum: [agentd, ssh] }
            desiredGeneration: { type: integer, format: int64 }
            readOnly: { type: boolean }
            planDigest:
              type: string
              description: 确认对象是具体计划（FR-12.7）；SSH 路径 AwaitingConfirmation 时非空
          additionalProperties: true
        status:
          type: object
          properties:
            phase:
              type: string
              enum: [Pending, Running, AwaitingConfirmation, CancelRequested, Succeeded, Failed, Unknown]
            startedAt: { type: string, format: date-time }
            finishedAt: { type: string, format: date-time }
            terminalModifier:
              type: string
              enum: [Superseded, Cancelled, Skipped]
            planExpiresAt:
              type: string
              format: date-time
              description: 确认窗口截止（§9.3 默认 30 分钟）；UI 显示"等待确认（剩余 T）"
            verify:
              type: object
              description: 门禁证据（FR-15.5）；UI 只呈现结论
              properties:
                desiredProjectionDigest: { type: string }
                observedProjectionDigest: { type: string }
                canonicalizationVersion: { type: string }
                inventorySeq: { type: integer, format: int64 }
                adapterHealth: { type: string, enum: [passed, failed, skipped] }
              additionalProperties: true
          additionalProperties: true
...
    AgentProfile:
      type: object
      required: [metadata]
      properties:
        metadata:
          $ref: "#/components/schemas/ObjectMeta"
        spec:
          type: object
          description: AgentProfile 规格（JSON 对象透传存储）
          additionalProperties: true
        status:
          type: object
          description: 无控制器，恒为 {}
          additionalProperties: true
    AgentProfileList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/AgentProfile"